feat: implement device selector component
