+++ packages/ui/src/components/cms/page-builder/PageToolbar.tsx
@@ -1,18 +1,8 @@
 import type { Locale } from "@/i18n/locales";
-import { DesktopIcon, LaptopIcon, MobileIcon } from "@radix-ui/react-icons";
-import {
-  Button,
-  Input,
-  Select,
-  SelectTrigger,
-  SelectValue,
-  SelectContent,
-  SelectItem,
-} from "../../atoms/shadcn";
-import { devicePresets, getLegacyPreset } from "@ui/utils/devicePresets";
+import { Button, Input } from "../../atoms/shadcn";
+import DeviceSelector from "../../common/DeviceSelector";
 
 interface Props {
-  viewport: "desktop" | "tablet" | "mobile";
   deviceId: string;
   setDeviceId: (id: string) => void;
   locale: Locale;
@@ -27,7 +17,6 @@ interface Props {
 }
 
 const PageToolbar = ({
-  viewport,
   deviceId,
   setDeviceId,
   locale,
@@ -41,38 +30,7 @@ const PageToolbar = ({
   setGridCols,
 }: Props) => (
   <div className="flex flex-col gap-4">
-    <div className="flex justify-end gap-2">
-      {(["desktop", "tablet", "mobile"] as const).map((t) => {
-        const preset = getLegacyPreset(t);
-        const Icon =
-          t === "desktop" ? DesktopIcon : t === "tablet" ? LaptopIcon : MobileIcon;
-        return (
-          <Button
-            key={t}
-            variant={deviceId === preset.id ? "default" : "outline"}
-            onClick={() => setDeviceId(preset.id)}
-            aria-label={t}
-          >
-            <Icon />
-            <span className="sr-only">
-              {t.charAt(0).toUpperCase() + t.slice(1)}
-            </span>
-          </Button>
-        );
-      })}
-      <Select value={deviceId} onValueChange={setDeviceId}>
-        <SelectTrigger aria-label="Device" className="w-40">
-          <SelectValue />
-        </SelectTrigger>
-        <SelectContent>
-          {devicePresets.map((p) => (
-            <SelectItem key={p.id} value={p.id}>
-              {p.label}
-            </SelectItem>
-          ))}
-        </SelectContent>
-      </Select>
-    </div>
+    <DeviceSelector deviceId={deviceId} onChange={setDeviceId} />
     <div className="flex items-center justify-end gap-2">
       <Button
         variant={showGrid ? "default" : "outline"}
